+++ src/WeihanLi.EntityFramework/IEFRepository.cs
@@ -15,6 +15,63 @@ namespace WeihanLi.EntityFramework
         where TEntity : class
     {
         /// <summary>
+        /// Delete a entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <returns>affected rows</returns>
+        int Delete(TEntity entity);
+
+        /// <summary>
+        /// Delete a entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <param name="cancellationToken">cancellationToken</param>
+        /// <returns>affected rows</returns>
+        Task<int> DeleteAsync(TEntity entity, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Update entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <param name="propertyExpressions">properties to update</param>
+        /// <returns>affected rows</returns>
+        int Update(TEntity entity, params Expression<Func<TEntity, object>>[] propertyExpressions);
+
+        /// <summary>
+        /// Update entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <param name="propertyExpressions">properties not to update</param>
+        /// <returns>affected rows</returns>
+        int UpdateWithout(TEntity entity, params Expression<Func<TEntity, object>>[] propertyExpressions);
+
+        /// <summary>
+        /// Update entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <param name="propertyExpressions">properties to update</param>
+        /// <param name="cancellationToken">cancellationToken</param>
+        /// <returns>affected rows</returns>
+        Task<int> UpdateAsync(TEntity entity, Expression<Func<TEntity, object>>[] propertyExpressions, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Update entity
+        /// </summary>
+        /// <param name="entity">entity</param>
+        /// <param name="propertyExpressions">properties not to update</param>
+        /// <param name="cancellationToken">cancellationToken</param>
+        /// <returns>affected rows</returns>
+        Task<int> UpdateWithoutAsync(TEntity entity, Expression<Func<TEntity, object>>[] propertyExpressions, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Finds an entity with the given primary key values. If found, is attached to the context and returned. If no entity is found, then null is returned.
+        /// </summary>
+        /// <param name="keyValues">The values of the primary key for the entity to be found.</param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
+        /// <returns>A <see cref="Task{TEntity}"/> that represents the asynchronous find operation. The task result contains the found entity or null.</returns>
+        Task<TEntity> FindAsync(object[] keyValues, CancellationToken cancellationToken = default);
+
+        /// <summary>
         /// Gets the <see cref="List{TEntity}"/> based on a predicate
         /// </summary>
         /// <param name="queryBuilderAction">queryBuilderAction</param>
@@ -27,7 +84,7 @@ namespace WeihanLi.EntityFramework
         /// <param name="selector"></param>
         /// <param name="queryBuilderAction">queryBuilderAction</param>
         /// <remarks>This method default no-tracking query.</remarks>
-        List<TResult> Get<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null);
+        List<TResult> GetResult<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null);
 
         /// <summary>
         /// Gets the <see cref="List{TEntity}"/> based on a predicate
@@ -44,7 +101,54 @@ namespace WeihanLi.EntityFramework
         /// <param name="queryBuilderAction">queryBuilder</param>
         /// <param name="cancellationToken"></param>
         /// <remarks>This method default no-tracking query.</remarks>
-        Task<List<TResult>> GetAsync<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, CancellationToken cancellationToken = default);
+        Task<List<TResult>> GetResultAsync<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Gets the <see cref="List{TEntity}"/> based on a predicate
+        /// </summary>
+        /// <param name="queryBuilderAction">queryBuilderAction</param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        bool Any(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null);
+
+        /// <summary>
+        /// Gets the <see cref="List{TEntity}"/> based on a predicate
+        /// </summary>
+        /// <param name="queryBuilderAction">queryBuilder</param>
+        /// <param name="cancellationToken">cancellationToken</param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        Task<bool> AnyAsync(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Gets the <see cref="List{TEntity}"/> based on a predicate
+        /// </summary>
+        /// <param name="queryBuilderAction">queryBuilderAction</param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        TEntity FirstOrDefault(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null);
+
+        /// <summary>
+        /// Gets the <see cref="List{TEntity}"/> based on a predicate
+        /// </summary>
+        /// <param name="selector"></param>
+        /// <param name="queryBuilderAction">queryBuilderAction</param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        TResult FirstOrDefaultResult<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null);
+
+        /// <summary>
+        /// Gets the <see cref="List{TEntity}"/> based on a predicate
+        /// </summary>
+        /// <param name="queryBuilderAction">queryBuilder</param>
+        /// <param name="cancellationToken">cancellationToken</param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        Task<TEntity> FirstOrDefaultAsync(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Gets the <see cref="List{TResult}"/> based on a predicate
+        /// </summary>
+        /// <param name="selector">selector</param>
+        /// <param name="queryBuilderAction">queryBuilder</param>
+        /// <param name="cancellationToken"></param>
+        /// <remarks>This method default no-tracking query.</remarks>
+        Task<TResult> FirstOrDefaultResultAsync<TResult>(Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Gets the <see cref="IPagedListModel{TEntity}"/> based on a predicate, orderby delegate and page information. This method default no-tracking query.
@@ -54,7 +158,7 @@ namespace WeihanLi.EntityFramework
         /// <param name="pageSize">The size of the page.</param>
         /// <returns>An <see cref="IPagedListModel{TEntity}"/> that contains elements that satisfy the condition specified by <paramref name="queryBuilderAction"/>.</returns>
         /// <remarks>This method default no-tracking query.</remarks>
-        IPagedListModel<TEntity> Paged(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, int pageNumber = 1, int pageSize = 20);
+        IPagedListModel<TEntity> GetPagedList(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, int pageNumber = 1, int pageSize = 20);
 
         /// <summary>
         /// Gets the <see cref="IPagedListModel{TEntity}"/> based on a predicate, orderby delegate and page information. This method default no-tracking query.
@@ -67,7 +171,7 @@ namespace WeihanLi.EntityFramework
         /// </param>
         /// <returns>An <see cref="IPagedListModel{TEntity}"/> that contains elements that satisfy the condition specified by <paramref name="queryBuilderAction"/>.</returns>
         /// <remarks>This method default no-tracking query.</remarks>
-        Task<IPagedListModel<TEntity>> PagedAsync(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, int pageNumber = 1, int pageSize = 20, CancellationToken cancellationToken = default);
+        Task<IPagedListModel<TEntity>> GetPagedListAsync(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null, int pageNumber = 1, int pageSize = 20, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Gets the <see cref="IPagedListModel{TResult}"/> based on a predicate, orderby delegate and page information. This method default no-tracking query.
@@ -78,8 +182,8 @@ namespace WeihanLi.EntityFramework
         /// <param name="pageSize">pageSize</param>
         /// <returns>An <see cref="IPagedListModel{TResult}"/> that contains elements that satisfy the condition specified by <paramref name="queryBuilderAction"/>.</returns>
         /// <remarks>This method default no-tracking query.</remarks>
-        IPagedListModel<TResult> Paged<TResult>([NotNull]Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null,
-                                                  int pageNumber = 1, int pageSize = 20) where TResult : class;
+        IPagedListModel<TResult> GetPagedListResult<TResult>([NotNull]Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null,
+                                                  int pageNumber = 1, int pageSize = 20);
 
         /// <summary>
         /// Gets the <see cref="IPagedListModel{TEntity}"/> based on a predicate, orderby delegate and page information. This method default no-tracking query.
@@ -93,100 +197,7 @@ namespace WeihanLi.EntityFramework
         /// <param name="pageNumber"></param>
         /// <returns>An <see cref="IPagedListModel{TEntity}"/> that contains elements that satisfy the condition specified by <paramref name="queryBuilderAction"/>.</returns>
         /// <remarks>This method default no-tracking query.</remarks>
-        Task<IPagedListModel<TResult>> PagedAsync<TResult>([NotNull]Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null,
-                                                             int pageNumber = 1, int pageSize = 20, CancellationToken cancellationToken = default) where TResult : class;
-
-        /// <summary>
-        /// Finds an entity with the given primary key values. If found, is attached to the context and returned. If no entity is found, then null is returned.
-        /// </summary>
-        /// <param name="keyValues">The values of the primary key for the entity to be found.</param>
-        /// <returns>The found entity or null.</returns>
-        TEntity Find(params object[] keyValues);
-
-        /// <summary>
-        /// Finds an entity with the given primary key values. If found, is attached to the context and returned. If no entity is found, then null is returned.
-        /// </summary>
-        /// <param name="keyValues">The values of the primary key for the entity to be found.</param>
-        /// <returns>A <see cref="Task{TEntity}"/> that represents the asynchronous find operation. The task result contains the found entity or null.</returns>
-        Task<TEntity> FindAsync(params object[] keyValues);
-
-        /// <summary>
-        /// Finds an entity with the given primary key values. If found, is attached to the context and returned. If no entity is found, then null is returned.
-        /// </summary>
-        /// <param name="keyValues">The values of the primary key for the entity to be found.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-        /// <returns>A <see cref="Task{TEntity}"/> that represents the asynchronous find operation. The task result contains the found entity or null.</returns>
-        Task<TEntity> FindAsync(object[] keyValues, CancellationToken cancellationToken);
-
-        int Update<TProperty>(TEntity entity, Expression<Func<TEntity, TProperty>> propertyExpression);
-
-        int Update<TProperty1, TProperty2>(TEntity entity, Expression<Func<TEntity, TProperty1>> propertyExpression1, Expression<Func<TEntity, TProperty2>> propertyExpression2);
-
-        int Update<TProperty1, TProperty2, TProperty3>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3);
-
-        int Update<TProperty1, TProperty2, TProperty3, TProperty4>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4
-            );
-
-        int Update<TProperty1, TProperty2, TProperty3, TProperty4, TProperty5>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4,
-            Expression<Func<TEntity, TProperty5>> propertyExpression5
-            );
-
-        int Update<TProperty1, TProperty2, TProperty3, TProperty4, TProperty5, TProperty6>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4,
-            Expression<Func<TEntity, TProperty5>> propertyExpression5,
-            Expression<Func<TEntity, TProperty6>> propertyExpression6
-            );
-
-        Task<int> UpdateAsync<TProperty>(TEntity entity, Expression<Func<TEntity, TProperty>> propertyExpression, CancellationToken cancellationToken = default);
-
-        Task<int> UpdateAsync<TProperty1, TProperty2>(TEntity entity, Expression<Func<TEntity, TProperty1>> propertyExpression1, Expression<Func<TEntity, TProperty2>> propertyExpression2, CancellationToken cancellationToken = default);
-
-        Task<int> UpdateAsync<TProperty1, TProperty2, TProperty3>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3, CancellationToken cancellationToken = default
-            );
-
-        Task<int> UpdateAsync<TProperty1, TProperty2, TProperty3, TProperty4>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4, CancellationToken cancellationToken = default
-            );
-
-        Task<int> UpdateAsync<TProperty1, TProperty2, TProperty3, TProperty4, TProperty5>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4,
-            Expression<Func<TEntity, TProperty5>> propertyExpression5, CancellationToken cancellationToken = default
-            );
-
-        Task<int> UpdateAsync<TProperty1, TProperty2, TProperty3, TProperty4, TProperty5, TProperty6>(TEntity entity,
-            Expression<Func<TEntity, TProperty1>> propertyExpression1,
-            Expression<Func<TEntity, TProperty2>> propertyExpression2,
-            Expression<Func<TEntity, TProperty3>> propertyExpression3,
-            Expression<Func<TEntity, TProperty4>> propertyExpression4,
-            Expression<Func<TEntity, TProperty5>> propertyExpression5,
-            Expression<Func<TEntity, TProperty6>> propertyExpression6, CancellationToken cancellationToken = default
-            );
-
-        int Update(TEntity entity, params string[] parameters);
-
-        Task<int> UpdateAsync(TEntity entity, CancellationToken cancellationToken = default, params string[] parameters);
+        Task<IPagedListModel<TResult>> GetPagedListResultAsync<TResult>([NotNull]Expression<Func<TEntity, TResult>> selector, Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null,
+                                                             int pageNumber = 1, int pageSize = 20, CancellationToken cancellationToken = default);
     }
 }