+++ src/WeihanLi.EntityFramework/IEFRepositoryBuilder.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace WeihanLi.EntityFramework
+{
+    public interface IEFRepositoryBuilder
+    {
+        IServiceCollection Services { get; }
+    }
+
+    internal class EFRepositoryBuilder : IEFRepositoryBuilder
+    {
+        public IServiceCollection Services { get; }
+
+        public EFRepositoryBuilder(IServiceCollection services) => Services = services;
+    }
+}