+++ docs/ReleaseNotes.md
@@ -1,12 +1,30 @@
 # Release Notes
 
-## 1.0.5
+## WeihanLi.EntityFramework
+
+
+### [WeihanLi.EntityFramework 1.0.8](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.8)
+
+- rename Get with selector method name => `GetResult`/`FirstOrDefaultResult`/`GetPagedListResult`
+
+
+### [WeihanLi.EntityFramework 1.0.7](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.7)
+
+- update `Update` for `EFRepository`
+- add `UpdateWithout`
+- remove none `QueryBuilder` extensions
+
+### [WeihanLi.EntityFramework 1.0.6](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.6)
+
+- add `FirstOrDefault` for `EFRepository`
+
+### [WeihanLi.EntityFramework 1.0.5](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.5)
 
 - add `CancellationToken` support for async operations
 - add `IEFRepositoryGenerator`
-- add `EFRepoQueryBuilder` for `EFRepository`
+- add `EFRepositoryQueryBuilder` for `EFRepository`
 - add [`sourceLink`](https://github.com/dotnet/sourcelink) support
 
-## 1.0.3
+### [WeihanLi.EntityFramework 1.0.3](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.3)
 
 - add `EFRepositoryGenerator`